pyproject.toml
[tool.poetry]
name = "budget-buddy"
version = "0.1.0"
description = ""
authors = ["web-lizzard <czarnous.adrian@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.*"
fastapi = "^0.104.1"
python-dotenv = "^1.0.0"
uvicorn = "^0.24.0.post1"
python-dateutil = "^2.8.2"
alembic = "^1.13.0"
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
black = "^23.11.0"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
